Bookの型定義を別ファイルに
