+++ src/service/aozora/types.ts
@@ -1,0 +1,9 @@
+export interface Book {
+  authorId: number;
+  author: string;
+  bookId: number;
+  book: string;
+  translator: string;
+  status: string;
+  since: string;
+}
